save contig name -> 2bit file name lookup; TODO: not tested; ucsc accession is not supported;
